pyproject.toml
[tool.poetry]
name = "guhema"
version = "0.1.0"
description = ""
authors = ["Christian Meter <cmeter@googlemail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.1.5"
django-modeltranslation = "^0.18.7"
django-autoslug = "^1.9.8"
easy-thumbnails = "^2.8.5"
django-pagedown = "^2.2.1"
django-forms-bootstrap = "^3.1.0"
django-markdownify = "^0.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
